* commit '1affff9db4aaa29ab66f9f5db76f575eeabd86b5':
  Fixed a clang-3.5 missing-prototype warning and added doxygen documentation for switch_channel_set_presence_data_vals. #doxygen
  Fixed dead code.
  Fixed trucation of value warning.
  Removed a useless called to abs.
  Removed an autological-pointer-compare from src/switch_utils.c.
  Fixed trucation of value warning.

+++ src/switch_core_media.c
@@ -7655,7 +7655,7 @@ SWITCH_DECLARE(switch_status_t) switch_core_media_receive_message(switch_core_se
 							maxlen = atol(p);
 							if ((q = strchr(p, ':'))) {
 								q++;
-								max_drift = abs(atol(q));
+								max_drift = abs(atoi(q));
 							}
 						}
 					}
